requirements.txt
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
coveralls==3.1.0
docopt==0.6.2
filestack-python==3.4.0
fpdf==1.7.2
httmock==1.4.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.0
python-barcode==0.13.1
requests==2.25.1
responses==0.13.3
six==1.16.0
toml==0.10.2
trafaret==2.0.2
urllib3==1.26.5
